popups

+++ Veto/src/main/java/pt4p1ae1/veto/GestionStock/dossierProduitController.java
@@ -3,7 +3,10 @@ package pt4p1ae1.veto.GestionStock;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.control.TextArea;
@@ -97,8 +100,17 @@ public class dossierProduitController extends ControllerSample implements Initia
 
     @FXML
     private void supprimerProduit() throws IOException {
-        Utils.PRODUIT_DAO.delete(currentProduit);
-        super.creatBtn("/fxml/rechercheProduit.fxml", (Stage) supprimer.getScene().getWindow());
+        Parent root = FXMLLoader.load(this.getClass().getResource("/fxml/popup.fxml"));
+        Scene scene = new Scene(root);
+        Stage stage = new Stage();
+        stage.setScene(scene);
+        stage.setTitle("Confirmation");
+        stage.showAndWait();
+
+        if(Utils.getConfirmation()) {
+            Utils.PRODUIT_DAO.delete(currentProduit);
+            super.creatBtn("/fxml/rechercheProduit.fxml", (Stage) supprimer.getScene().getWindow());
+        }
     }
 
     public void display(){
